Rename concept.schema to concept.logic
